fixed potential bug at high depths

+++ src/tables.rs
@@ -43,7 +43,7 @@ pub fn tt_clear() {
 /// 1. Prioritise replacing entries for the same position (key) that have lower depth.
 /// 2. Fill empty entries in bucket.
 /// 3. Replace lowest depth entry in bucket.
-pub fn tt_push(zobrist: u64, best_move: u16, depth: i8, bound: u8, mut score: i16, ply: i8) {
+pub fn tt_push(zobrist: u64, best_move: u16, depth: i8, bound: u8, mut score: i16, ply: i16) {
     let key: u16 = (zobrist >> 48) as u16;
     let idx: usize = (zobrist as usize) & (unsafe{TT_SIZE} - 1);
     let bucket: &mut [HashEntry] = unsafe{&mut TT[idx]};
@@ -60,25 +60,25 @@ pub fn tt_push(zobrist: u64, best_move: u16, depth: i8, bound: u8, mut score: i1
             continue;
         }
     }
-    score += if score > MATE_THRESHOLD {ply} else if score < -MATE_THRESHOLD {-ply} else {0} as i16;
+    score += if score > MATE_THRESHOLD {ply} else if score < -MATE_THRESHOLD {-ply} else {0};
     bucket[desired_idx] = HashEntry {key, best_move, depth, bound, score };
 }
 
-pub fn tt_probe(zobrist: u64, ply: i8) -> Option<HashEntry> {
+pub fn tt_probe(zobrist: u64, ply: i16) -> Option<HashEntry> {
     let key: u16 = (zobrist >> 48) as u16;
     let idx: usize = (zobrist as usize) & (unsafe{TT_SIZE} - 1);
     let bucket: &[HashEntry; 8] = unsafe{&TT[idx]};
     for entry in bucket {
         if entry.key == key {
             let mut res: HashEntry = *entry;
-            res.score += if res.score > MATE_THRESHOLD {-ply} else if res.score < -MATE_THRESHOLD {ply} else {0} as i16;
+            res.score += if res.score > MATE_THRESHOLD {-ply} else if res.score < -MATE_THRESHOLD {ply} else {0};
             return Some(res);
         }
     }
     None
 }
 
-pub fn kt_push(m: u16, p: i8) {
+pub fn kt_push(m: u16, p: i16) {
     unsafe {
         let ply: usize = p as usize - 1;
         let new: u16 = if KT[ply].contains(&m) {KT[ply][2]} else {m};
@@ -93,6 +93,6 @@ pub fn kt_clear() {
 }
 
 #[inline(always)]
-pub fn kt_get(ply: i8) -> [u16; 3] {
+pub fn kt_get(ply: i16) -> [u16; 3] {
     unsafe{KT[ply as usize]}
 }
